# artifacts/openapi-management.yaml
openapi: 3.1.0
info:
  title: Cutctx Management API
  description: |
    Management and observability endpoints for the Cutctx context compression proxy.
    
    These endpoints are served by the Python management layer alongside the Rust
    reverse proxy. They require admin API key authentication when configured.
    
    **Authentication:** When `CUTCTX_ADMIN_API_KEY` is set, all management
    endpoints require either:
    - `Authorization: Bearer <admin-api-key>` header
    - `X-Cutctx-Admin-Key: <admin-api-key>` header
    
    When no admin API key is configured, endpoints are open (backward-compatible).
  version: "0.26.0"
  contact:
    name: Cutctx Labs
    email: hello@cutctx.dev
    url: https://cutctx.dev
  license:
    name: Cutctx Commercial License
    url: https://pitchtoship.com/legal/commercial-license
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://pitchtoship.com
    description: PitchToShip license and billing backend
security:
  - BearerAuth: []
  - AdminKeyHeader: []
...
paths:
...
  /v1/license/activate:
    post:
      tags: [Admin]
      summary: Activate a license instance
      description: Register a proxy instance activation against a license key.
      operationId: activateLicense
      security: []
      responses:
        "200":
          description: Activated
        "401":
          description: Invalid license
        "403":
          description: Revoked or expired
...
  /v1/license/checkout-seat:
    post:
      tags: [Admin]
      summary: Checkout a seat lease
      description: Lease a seat for a user.
      operationId: checkoutSeat
      security: []
      responses:
        "200":
          description: Seat leased
        "429":
          description: No seats available
  /v1/license/start-trial:
    post:
      tags: [Admin]
      summary: Start a trial
      description: Start a 14-day server-side trial.
      operationId: startTrial
      security: []
      responses:
        "200":
          description: Trial started
        "409":
          description: Trial already exists
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin API key as Bearer token
    AdminKeyHeader:
      type: apiKey
      in: header
      name: X-Cutctx-Admin-Key
      description: Admin API key in custom header
...